+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/AsyncKafkaConsumerTest.java
@@ -16,125 +16,621 @@
  */
 package org.apache.kafka.clients.consumer.internals;
 
+import static java.util.Arrays.asList;
 import static java.util.Collections.singleton;
+import static java.util.Collections.singletonList;
+import static org.apache.kafka.clients.consumer.internals.ConsumerUtils.THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED;
+import static org.apache.kafka.common.utils.Utils.mkEntry;
+import static org.apache.kafka.common.utils.Utils.mkMap;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeast;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.time.Duration;
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Properties;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Future;
-import org.apache.kafka.clients.ClientResponse;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.apache.kafka.clients.Metadata.LeaderAndEpoch;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.consumer.ConsumerGroupMetadata;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.consumer.ConsumerRecords;
+import org.apache.kafka.clients.consumer.GroupProtocol;
 import org.apache.kafka.clients.consumer.OffsetAndMetadata;
-import org.apache.kafka.clients.consumer.internals.events.ApplicationEvent;
+import org.apache.kafka.clients.consumer.OffsetAndTimestamp;
+import org.apache.kafka.clients.consumer.OffsetCommitCallback;
 import org.apache.kafka.clients.consumer.internals.events.ApplicationEventHandler;
+import org.apache.kafka.clients.consumer.internals.events.AssignmentChangeApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.BackgroundEvent;
+import org.apache.kafka.clients.consumer.internals.events.CommitApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.CompletableApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.ErrorBackgroundEvent;
 import org.apache.kafka.clients.consumer.internals.events.FetchCommittedOffsetsApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.GroupMetadataUpdateEvent;
+import org.apache.kafka.clients.consumer.internals.events.ListOffsetsApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.NewTopicsMetadataUpdateRequestEvent;
 import org.apache.kafka.clients.consumer.internals.events.ResetPositionsApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.SubscriptionChangeApplicationEvent;
+import org.apache.kafka.clients.consumer.internals.events.UnsubscribeApplicationEvent;
 import org.apache.kafka.clients.consumer.internals.events.ValidatePositionsApplicationEvent;
-import org.apache.kafka.common.Node;
+import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.TopicPartition;
+import org.apache.kafka.common.errors.GroupAuthorizationException;
+import org.apache.kafka.common.errors.InvalidGroupIdException;
+import org.apache.kafka.common.errors.TimeoutException;
 import org.apache.kafka.common.errors.WakeupException;
-import org.apache.kafka.common.message.OffsetCommitResponseData;
-import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.Errors;
-import org.apache.kafka.common.requests.FindCoordinatorResponse;
-import org.apache.kafka.common.requests.OffsetCommitResponse;
-import org.apache.kafka.common.requests.RequestHeader;
+import org.apache.kafka.common.requests.JoinGroupRequest;
+import org.apache.kafka.common.requests.ListOffsetsRequest;
+import org.apache.kafka.common.serialization.StringDeserializer;
+import org.apache.kafka.common.utils.MockTime;
+import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.common.utils.Timer;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatchers;
-import org.mockito.MockedConstruction;
-import org.mockito.stubbing.Answer;
+import org.mockito.Mockito;
 
+@SuppressWarnings("unchecked")
 public class AsyncKafkaConsumerTest {
 
-    private AsyncKafkaConsumer<?, ?> consumer;
-    private ConsumerTestBuilder.AsyncKafkaConsumerTestBuilder testBuilder;
-    private ApplicationEventHandler applicationEventHandler;
+    private AsyncKafkaConsumer<String, String> consumer = null;
 
-    @BeforeEach
-    public void setup() {
-        // By default, the consumer is part of a group and autoCommit is enabled.
-        setup(ConsumerTestBuilder.createDefaultGroupInformation(), true);
-    }
-
-    private void setup(Optional<ConsumerTestBuilder.GroupInformation> groupInfo, boolean enableAutoCommit) {
-        testBuilder = new ConsumerTestBuilder.AsyncKafkaConsumerTestBuilder(groupInfo, enableAutoCommit, true);
-        applicationEventHandler = testBuilder.applicationEventHandler;
-        consumer = testBuilder.consumer;
-    }
+    private final Time time = new MockTime(1);
+    private final FetchCollector<String, String> fetchCollector = mock(FetchCollector.class);
+    private final ApplicationEventHandler applicationEventHandler = mock(ApplicationEventHandler.class);
+    private final ConsumerMetadata metadata = mock(ConsumerMetadata.class);
+    private final LinkedBlockingQueue<BackgroundEvent> backgroundEventQueue = new LinkedBlockingQueue<>();
 
     @AfterEach
-    public void cleanup() {
-        if (testBuilder != null) {
-            shutDown();
+    public void resetAll() {
+        backgroundEventQueue.clear();
+        if (consumer != null) {
+            consumer.close();
         }
+        consumer = null;
+        Mockito.framework().clearInlineMocks();
     }
 
-    private void shutDown() {
-        prepAutocommitOnClose();
-        testBuilder.close();
+    private AsyncKafkaConsumer<String, String> setup() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, "group-id");
+        final ConsumerConfig config = new ConsumerConfig(props);
+        return setup(config);
     }
 
-    private void resetWithEmptyGroupId() {
-        // Create a consumer that is not configured as part of a group.
-        cleanup();
-        setup(Optional.empty(), false);
+    private AsyncKafkaConsumer<String, String> setupWithoutGroupId() {
+        final Properties props = requiredConsumerProperties();
+        final ConsumerConfig config = new ConsumerConfig(props);
+        return setup(config);
     }
 
-    /**
-     * This is a rather ugly bit of code. Not my choice :(
-     *
-     * <p/>
-     *
-     * Inside the {@link org.apache.kafka.clients.consumer.Consumer#committed(Set, Duration)} call we create an
-     * instance of {@link FetchCommittedOffsetsApplicationEvent} that holds the partitions and internally holds a
-     * {@link CompletableFuture}. We want to test different behaviours of the {@link Future#get()}, such as
-     * returning normally, timing out, throwing an error, etc. By mocking the construction of the event object that
-     * is created, we can affect that behavior.
-     */
-    private static MockedConstruction<FetchCommittedOffsetsApplicationEvent> offsetFetchEventMocker(CompletableFuture<Map<TopicPartition, OffsetAndMetadata>> future) {
-        // This "answer" is where we pass the future to be invoked by the ConsumerUtils.getResult() method
-        Answer<Map<TopicPartition, OffsetAndMetadata>> getInvocationAnswer = invocation -> {
-            // This argument captures the actual argument value that was passed to the event's get() method, so we
-            // just "forward" that value to our mocked call
-            Timer timer = invocation.getArgument(0);
-            return ConsumerUtils.getResult(future, timer);
-        };
+    @SuppressWarnings("UnusedReturnValue")
+    private AsyncKafkaConsumer<String, String> setupWithEmptyGroupId() {
+        final Properties props = requiredConsumerPropertiesAndGroupId("");
+        final ConsumerConfig config = new ConsumerConfig(props);
+        return setup(config);
+    }
 
-        MockedConstruction.MockInitializer<FetchCommittedOffsetsApplicationEvent> mockInitializer = (mock, ctx) -> {
-            // When the event's get() method is invoked, we call the "answer" method just above
-            when(mock.get(any())).thenAnswer(getInvocationAnswer);
-
-            // When the event's type() method is invoked, we have to return the type as it will be null in the mock
-            when(mock.type()).thenReturn(ApplicationEvent.Type.FETCH_COMMITTED_OFFSETS);
-
-            // This is needed for the WakeupTrigger code that keeps track of the active task
-            when(mock.future()).thenReturn(future);
-        };
-
-        return mockConstruction(FetchCommittedOffsetsApplicationEvent.class, mockInitializer);
+    private AsyncKafkaConsumer<String, String> setup(ConsumerConfig config) {
+        return new AsyncKafkaConsumer<>(
+            config,
+            new StringDeserializer(),
+            new StringDeserializer(),
+            time,
+            (a, b, c, d, e, f) -> applicationEventHandler,
+            (a, b, c, d, e, f, g) -> fetchCollector,
+            (a, b, c, d) -> metadata,
+            backgroundEventQueue
+        );
     }
 
     @Test
+    public void testSuccessfulStartupShutdown() {
+        consumer = setup();
+        assertDoesNotThrow(() -> consumer.close());
+    }
+
+    @Test
+    public void testInvalidGroupId() {
+        KafkaException e = assertThrows(KafkaException.class, this::setupWithEmptyGroupId);
+        assertTrue(e.getCause() instanceof InvalidGroupIdException);
+    }
+
+    @Test
+    public void testCommitAsyncWithNullCallback() {
+        consumer = setup();
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, OffsetAndMetadata> offsets = new HashMap<>();
+        offsets.put(t0, new OffsetAndMetadata(10L));
+        offsets.put(t1, new OffsetAndMetadata(20L));
+
+        consumer.commitAsync(offsets, null);
+
+        final ArgumentCaptor<CommitApplicationEvent> commitEventCaptor = ArgumentCaptor.forClass(CommitApplicationEvent.class);
+        verify(applicationEventHandler).add(commitEventCaptor.capture());
+        final CommitApplicationEvent commitEvent = commitEventCaptor.getValue();
+        assertEquals(offsets, commitEvent.offsets());
+        assertDoesNotThrow(() -> commitEvent.future().complete(null));
+        assertDoesNotThrow(() -> consumer.commitAsync(offsets, null));
+    }
+
+    private static Stream<Exception> commitExceptionSupplier() {
+        return Stream.of(
+            null,  // For the successful completion scenario
+            new KafkaException("Test exception"),
+            new GroupAuthorizationException("Group authorization exception"));
+    }
+
+    @ParameterizedTest
+    @MethodSource("commitExceptionSupplier")
+    public void testCommitAsyncUserSuppliedCallback(Exception exception) {
+        consumer = setup();
+
+        Map<TopicPartition, OffsetAndMetadata> offsets = new HashMap<>();
+        offsets.put(new TopicPartition("my-topic", 1), new OffsetAndMetadata(200L));
+
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof CommitApplicationEvent);
+            if (exception == null) {
+                event.future().complete(null);
+            } else {
+                event.future().completeExceptionally(exception);
+            }
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        MockCommitCallback callback = new MockCommitCallback();
+        assertDoesNotThrow(() -> consumer.commitAsync(offsets, callback));
+        consumer.maybeInvokeCommitCallbacks();
+
+        if (exception == null) {
+            assertNull(callback.exception);
+        } else {
+            assertSame(exception.getClass(), callback.exception.getClass());
+        }
+    }
+
+    @Test
+    public void testFencedInstanceException() {
+        consumer = setup();
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof CommitApplicationEvent);
+            event.future().completeExceptionally(Errors.FENCED_INSTANCE_ID.exception());
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        assertDoesNotThrow(() -> consumer.commitAsync());
+    }
+
+    @Test
+    public void testCommitted() {
+        consumer = setup();
+        Map<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = mockTopicPartitionOffset();
+
+        when(applicationEventHandler.addAndGet(any(), any())).thenAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof FetchCommittedOffsetsApplicationEvent);
+            return topicPartitionOffsets;
+        });
+
+        assertDoesNotThrow(() -> consumer.committed(topicPartitionOffsets.keySet(), Duration.ofMillis(1000)));
+    }
+
+    @Test
+    public void testCommittedExceptionThrown() {
+        consumer = setup();
+        Map<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = mockTopicPartitionOffset();
+        when(applicationEventHandler.addAndGet(any(), any())).thenAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof FetchCommittedOffsetsApplicationEvent);
+            throw new KafkaException("Test exception");
+        });
+
+        assertThrows(KafkaException.class, () -> consumer.committed(topicPartitionOffsets.keySet(), Duration.ofMillis(1000)));
+    }
+
+    @Test
+    public void testCommittedLeaderEpochUpdate() {
+        consumer = setup();
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        final TopicPartition t2 = new TopicPartition("t0", 4);
+        HashMap<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = new HashMap<>();
+        topicPartitionOffsets.put(t0, new OffsetAndMetadata(10L, Optional.of(2), ""));
+        topicPartitionOffsets.put(t1, null);
+        topicPartitionOffsets.put(t2, new OffsetAndMetadata(20L, Optional.of(3), ""));
+        when(applicationEventHandler.addAndGet(any(), any())).thenAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof FetchCommittedOffsetsApplicationEvent);
+            return topicPartitionOffsets;
+        });
+
+        assertDoesNotThrow(() -> consumer.committed(topicPartitionOffsets.keySet(), Duration.ofMillis(1000)));
+
+        verify(metadata).updateLastSeenEpochIfNewer(t0, 2);
+        verify(metadata).updateLastSeenEpochIfNewer(t2, 3);
+    }
+
+
+    @Test
+    public void testWakeupBeforeCallingPoll() {
+        consumer = setup();
+        final String topicName = "foo";
+        final int partition = 3;
+        final TopicPartition tp = new TopicPartition(topicName, partition);
+        doReturn(Fetch.empty()).when(fetchCollector).collectFetch(any(FetchBuffer.class));
+        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
+        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(LeaderAndEpoch.noLeaderOrEpoch()).when(metadata).currentLeader(any());
+        consumer.assign(singleton(tp));
+
+        consumer.wakeup();
+
+        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ZERO));
+        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
+    }
+
+    @Test
+    public void testWakeupAfterEmptyFetch() {
+        consumer = setup();
+        final String topicName = "foo";
+        final int partition = 3;
+        final TopicPartition tp = new TopicPartition(topicName, partition);
+        doAnswer(invocation -> {
+            consumer.wakeup();
+            return Fetch.empty();
+        }).when(fetchCollector).collectFetch(any(FetchBuffer.class));
+        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
+        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(LeaderAndEpoch.noLeaderOrEpoch()).when(metadata).currentLeader(any());
+        consumer.assign(singleton(tp));
+
+        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ofMinutes(1)));
+        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
+    }
+
+    @Test
+    public void testWakeupAfterNonEmptyFetch() {
+        consumer = setup();
+        final String topicName = "foo";
+        final int partition = 3;
+        final TopicPartition tp = new TopicPartition(topicName, partition);
+        final List<ConsumerRecord<String, String>> records = asList(
+            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
+            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
+        );
+        doAnswer(invocation -> {
+            consumer.wakeup();
+            return Fetch.forPartition(tp, records, true);
+        }).when(fetchCollector).collectFetch(Mockito.any(FetchBuffer.class));
+        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
+        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(LeaderAndEpoch.noLeaderOrEpoch()).when(metadata).currentLeader(any());
+        consumer.assign(singleton(tp));
+
+        // since wakeup() is called when the non-empty fetch is returned the wakeup should be ignored
+        assertDoesNotThrow(() -> consumer.poll(Duration.ofMinutes(1)));
+        // the previously ignored wake-up should not be ignored in the next call
+        assertThrows(WakeupException.class, () -> consumer.poll(Duration.ZERO));
+    }
+
+    @Test
+    public void testClearWakeupTriggerAfterPoll() {
+        consumer = setup();
+        final String topicName = "foo";
+        final int partition = 3;
+        final TopicPartition tp = new TopicPartition(topicName, partition);
+        final List<ConsumerRecord<String, String>> records = asList(
+            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
+            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
+        );
+        doReturn(Fetch.forPartition(tp, records, true))
+            .when(fetchCollector).collectFetch(any(FetchBuffer.class));
+        Map<TopicPartition, OffsetAndMetadata> offsets = mkMap(mkEntry(tp, new OffsetAndMetadata(1)));
+        doReturn(offsets).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(LeaderAndEpoch.noLeaderOrEpoch()).when(metadata).currentLeader(any());
+        consumer.assign(singleton(tp));
+
+        consumer.poll(Duration.ZERO);
+
+        assertDoesNotThrow(() -> consumer.poll(Duration.ZERO));
+    }
+
+    @Test
+    public void testEnsureCallbackExecutedByApplicationThread() {
+        consumer = setup();
+        final String currentThread = Thread.currentThread().getName();
+        MockCommitCallback callback = new MockCommitCallback();
+
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof CommitApplicationEvent);
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        assertDoesNotThrow(() -> consumer.commitAsync(new HashMap<>(), callback));
+        assertEquals(1, consumer.callbacks());
+        consumer.maybeInvokeCommitCallbacks();
+        assertEquals(currentThread, callback.completionThread);
+    }
+
+
+    @Test
+    @SuppressWarnings("deprecation")
+    public void testPollLongThrowsException() {
+        consumer = setup();
+        Exception e = assertThrows(UnsupportedOperationException.class, () -> consumer.poll(0L));
+        assertEquals("Consumer.poll(long) is not supported when \"group.protocol\" is \"consumer\". " +
+            "This method is feprecated and will be removed in the next major release.", e.getMessage());
+    }
+
+    @Test
+    public void testFailOnClosedConsumer() {
+        consumer = setup();
+        consumer.close();
+        final IllegalStateException res = assertThrows(IllegalStateException.class, consumer::assignment);
+        assertEquals("This consumer has already been closed.", res.getMessage());
+    }
+
+    @Test
+    public void testCommitSyncLeaderEpochUpdate() {
+        consumer = setup();
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = new HashMap<>();
+        topicPartitionOffsets.put(t0, new OffsetAndMetadata(10L, Optional.of(2), ""));
+        topicPartitionOffsets.put(t1, new OffsetAndMetadata(20L, Optional.of(1), ""));
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof CommitApplicationEvent);
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        assertDoesNotThrow(() -> consumer.commitSync(topicPartitionOffsets));
+
+        verify(metadata).updateLastSeenEpochIfNewer(t0, 2);
+        verify(metadata).updateLastSeenEpochIfNewer(t1, 1);
+    }
+
+    @Test
+    public void testCommitAsyncLeaderEpochUpdate() {
+        consumer = setup();
+        OffsetCommitCallback callback = mock(OffsetCommitCallback.class);
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = new HashMap<>();
+        topicPartitionOffsets.put(t0, new OffsetAndMetadata(10L, Optional.of(2), ""));
+        topicPartitionOffsets.put(t1, new OffsetAndMetadata(20L, Optional.of(1), ""));
+        consumer.assign(Arrays.asList(t0, t1));
+
+        consumer.commitAsync(topicPartitionOffsets, callback);
+
+        verify(metadata).updateLastSeenEpochIfNewer(t0, 2);
+        verify(metadata).updateLastSeenEpochIfNewer(t1, 1);
+    }
+
+    private static class MockCommitCallback implements OffsetCommitCallback {
+        public int invoked = 0;
+        public Exception exception = null;
+        public String completionThread;
+
+        @Override
+        public void onComplete(Map<TopicPartition, OffsetAndMetadata> offsets, Exception exception) {
+            invoked++;
+            this.completionThread = Thread.currentThread().getName();
+            this.exception = exception;
+        }
+    }
+
+    @Test
+    public void testAssign() {
+        consumer = setup();
+        final TopicPartition tp = new TopicPartition("foo", 3);
+        consumer.assign(singleton(tp));
+        assertTrue(consumer.subscription().isEmpty());
+        assertTrue(consumer.assignment().contains(tp));
+        verify(applicationEventHandler).add(any(AssignmentChangeApplicationEvent.class));
+        verify(applicationEventHandler).add(any(NewTopicsMetadataUpdateRequestEvent.class));
+    }
+
+    @Test
+    public void testAssignOnNullTopicPartition() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class, () -> consumer.assign(null));
+    }
+
+    @Test
+    public void testAssignOnEmptyTopicPartition() {
+        consumer = setup();
+
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof UnsubscribeApplicationEvent);
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        consumer.assign(Collections.emptyList());
+        assertTrue(consumer.subscription().isEmpty());
+        assertTrue(consumer.assignment().isEmpty());
+    }
+
+    @Test
+    public void testAssignOnNullTopicInPartition() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class, () -> consumer.assign(singleton(new TopicPartition(null, 0))));
+    }
+
+    @Test
+    public void testAssignOnEmptyTopicInPartition() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class, () -> consumer.assign(singleton(new TopicPartition("  ", 0))));
+    }
+
+    @Test
+    public void testBeginningOffsetsFailsIfNullPartitions() {
+        consumer = setup();
+        assertThrows(NullPointerException.class, () -> consumer.beginningOffsets(null,
+            Duration.ofMillis(1)));
+    }
+
+    @Test
+    public void testBeginningOffsets() {
+        consumer = setup();
+        Map<TopicPartition, OffsetAndTimestamp> expectedOffsetsAndTimestamp =
+            mockOffsetAndTimestamp();
+        Set<TopicPartition> partitions = expectedOffsetsAndTimestamp.keySet();
+        doReturn(expectedOffsetsAndTimestamp).when(applicationEventHandler).addAndGet(any(), any());
+        Map<TopicPartition, Long> result =
+            assertDoesNotThrow(() -> consumer.beginningOffsets(partitions,
+                Duration.ofMillis(1)));
+        Map<TopicPartition, Long> expectedOffsets = expectedOffsetsAndTimestamp.entrySet().stream()
+            .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().offset()));
+        assertEquals(expectedOffsets, result);
+        verify(applicationEventHandler).addAndGet(ArgumentMatchers.isA(ListOffsetsApplicationEvent.class),
+            ArgumentMatchers.isA(Timer.class));
+    }
+
+    @Test
+    public void testBeginningOffsetsThrowsKafkaExceptionForUnderlyingExecutionFailure() {
+        consumer = setup();
+        Set<TopicPartition> partitions = mockTopicPartitionOffset().keySet();
+        Throwable eventProcessingFailure = new KafkaException("Unexpected failure " +
+            "processing List Offsets event");
+        doThrow(eventProcessingFailure).when(applicationEventHandler).addAndGet(any(), any());
+        Throwable consumerError = assertThrows(KafkaException.class,
+            () -> consumer.beginningOffsets(partitions,
+                Duration.ofMillis(1)));
+        assertEquals(eventProcessingFailure, consumerError);
+        verify(applicationEventHandler).addAndGet(ArgumentMatchers.isA(ListOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+    }
+
+    @Test
+    public void testBeginningOffsetsTimeoutOnEventProcessingTimeout() {
+        consumer = setup();
+        doThrow(new TimeoutException()).when(applicationEventHandler).addAndGet(any(), any());
+        assertThrows(TimeoutException.class,
+            () -> consumer.beginningOffsets(
+                Collections.singletonList(new TopicPartition("t1", 0)),
+                Duration.ofMillis(1)));
+        verify(applicationEventHandler).addAndGet(ArgumentMatchers.isA(ListOffsetsApplicationEvent.class),
+            ArgumentMatchers.isA(Timer.class));
+    }
+
+    @Test
+    public void testOffsetsForTimesOnNullPartitions() {
+        consumer = setup();
+        assertThrows(NullPointerException.class, () -> consumer.offsetsForTimes(null,
+            Duration.ofMillis(1)));
+    }
+
+    @Test
+    public void testOffsetsForTimesFailsOnNegativeTargetTimes() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class,
+            () -> consumer.offsetsForTimes(Collections.singletonMap(new TopicPartition(
+                    "topic1", 1), ListOffsetsRequest.EARLIEST_TIMESTAMP),
+                Duration.ofMillis(1)));
+
+        assertThrows(IllegalArgumentException.class,
+            () -> consumer.offsetsForTimes(Collections.singletonMap(new TopicPartition(
+                    "topic1", 1), ListOffsetsRequest.LATEST_TIMESTAMP),
+                Duration.ofMillis(1)));
+
+        assertThrows(IllegalArgumentException.class,
+            () -> consumer.offsetsForTimes(Collections.singletonMap(new TopicPartition(
+                    "topic1", 1), ListOffsetsRequest.MAX_TIMESTAMP),
+                Duration.ofMillis(1)));
+    }
+
+    @Test
+    public void testOffsetsForTimes() {
+        consumer = setup();
+        Map<TopicPartition, OffsetAndTimestamp> expectedResult = mockOffsetAndTimestamp();
+        Map<TopicPartition, Long> timestampToSearch = mockTimestampToSearch();
+
+        doReturn(expectedResult).when(applicationEventHandler).addAndGet(any(), any());
+        Map<TopicPartition, OffsetAndTimestamp> result =
+            assertDoesNotThrow(() -> consumer.offsetsForTimes(timestampToSearch, Duration.ofMillis(1)));
+        assertEquals(expectedResult, result);
+        verify(applicationEventHandler).addAndGet(ArgumentMatchers.isA(ListOffsetsApplicationEvent.class),
+            ArgumentMatchers.isA(Timer.class));
+    }
+
+    // This test ensures same behaviour as the current consumer when offsetsForTimes is called
+    // with 0 timeout. It should return map with all requested partitions as keys, with null
+    // OffsetAndTimestamp as value.
+    @Test
+    public void testOffsetsForTimesWithZeroTimeout() {
+        consumer = setup();
+        TopicPartition tp = new TopicPartition("topic1", 0);
+        Map<TopicPartition, OffsetAndTimestamp> expectedResult =
+            Collections.singletonMap(tp, null);
+        Map<TopicPartition, Long> timestampToSearch = Collections.singletonMap(tp, 5L);
+
+        Map<TopicPartition, OffsetAndTimestamp> result =
+            assertDoesNotThrow(() -> consumer.offsetsForTimes(timestampToSearch,
+                Duration.ofMillis(0)));
+        assertEquals(expectedResult, result);
+        verify(applicationEventHandler, never()).addAndGet(ArgumentMatchers.isA(ListOffsetsApplicationEvent.class),
+            ArgumentMatchers.isA(Timer.class));
+    }
+
+    @Test
+    public void testWakeupCommitted() {
+        consumer = setup();
+        final HashMap<TopicPartition, OffsetAndMetadata> offsets = mockTopicPartitionOffset();
+        doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            Timer timer = invocation.getArgument(1);
+            assertTrue(event instanceof FetchCommittedOffsetsApplicationEvent);
+            assertTrue(event.future().isCompletedExceptionally());
+            return ConsumerUtils.getResult(event.future(), timer);
+        })
+            .when(applicationEventHandler)
+            .addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+
+        consumer.wakeup();
+        assertThrows(WakeupException.class, () -> consumer.committed(offsets.keySet()));
+        assertNull(consumer.wakeupTrigger().getPendingTask());
+    }
+
+
+    @Test
     public void testRefreshCommittedOffsetsSuccess() {
+        consumer = setup();
         TopicPartition partition = new TopicPartition("t1", 1);
         Set<TopicPartition> partitions = Collections.singleton(partition);
         Map<TopicPartition, OffsetAndMetadata> committedOffsets = Collections.singletonMap(partition, new OffsetAndMetadata(10L));
@@ -143,6 +639,7 @@ public class AsyncKafkaConsumerTest {
 
     @Test
     public void testRefreshCommittedOffsetsSuccessButNoCommittedOffsetsFound() {
+        consumer = setup();
         TopicPartition partition = new TopicPartition("t1", 1);
         Set<TopicPartition> partitions = Collections.singleton(partition);
         Map<TopicPartition, OffsetAndMetadata> committedOffsets = Collections.emptyMap();
@@ -151,102 +648,403 @@ public class AsyncKafkaConsumerTest {
 
     @Test
     public void testRefreshCommittedOffsetsShouldNotResetIfFailedWithTimeout() {
+        consumer = setup();
         testUpdateFetchPositionsWithFetchCommittedOffsetsTimeout(true);
     }
 
     @Test
     public void testRefreshCommittedOffsetsNotCalledIfNoGroupId() {
+        consumer = setupWithoutGroupId();
         // Create consumer without group id so committed offsets are not used for updating positions
-        resetWithEmptyGroupId();
         testUpdateFetchPositionsWithFetchCommittedOffsetsTimeout(false);
     }
 
     private void testUpdateFetchPositionsWithFetchCommittedOffsetsTimeout(boolean committedOffsetsEnabled) {
-        // Uncompleted future that will time out if used
-        CompletableFuture<Map<TopicPartition, OffsetAndMetadata>> committedFuture = new CompletableFuture<>();
+        // Do not complete future that is passed into addAndGet, to timeout
+        doReturn(mkMap())
+            .when(applicationEventHandler)
+            .addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(Fetch.empty()).when(fetchCollector).collectFetch(any(FetchBuffer.class));
 
         consumer.assign(singleton(new TopicPartition("t1", 1)));
 
-        try (MockedConstruction<FetchCommittedOffsetsApplicationEvent> ignored = offsetFetchEventMocker(committedFuture)) {
-            // Poll with 250ms timeout to give the background thread time to process the events without timing out
-            consumer.poll(Duration.ofMillis(250));
+        // Poll with 250ms timeout to give the background thread time to process the events without timing out
+        consumer.poll(Duration.ofMillis(250));
 
+        verify(applicationEventHandler, atLeast(1))
+            .addAndGet(ArgumentMatchers.isA(ValidatePositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+
+        if (committedOffsetsEnabled) {
+            // Verify there was an FetchCommittedOffsets event and no ResetPositions event
             verify(applicationEventHandler, atLeast(1))
-                .addAndGet(ArgumentMatchers.isA(ValidatePositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-
-            if (committedOffsetsEnabled) {
-                // Verify there was an FetchCommittedOffsets event and no ResetPositions event
-                verify(applicationEventHandler, atLeast(1))
-                    .addAndGet(ArgumentMatchers.isA(FetchCommittedOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-                verify(applicationEventHandler, never())
-                    .addAndGet(ArgumentMatchers.isA(ResetPositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-            } else {
-                // Verify there was not any FetchCommittedOffsets event but there should be a ResetPositions
-                verify(applicationEventHandler, never())
-                    .addAndGet(ArgumentMatchers.isA(FetchCommittedOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-                verify(applicationEventHandler, atLeast(1))
-                    .addAndGet(ArgumentMatchers.isA(ResetPositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-            }
-        }
-    }
-
-    private void testRefreshCommittedOffsetsSuccess(Set<TopicPartition> partitions,
-                                                    Map<TopicPartition, OffsetAndMetadata> committedOffsets) {
-        CompletableFuture<Map<TopicPartition, OffsetAndMetadata>> committedFuture = new CompletableFuture<>();
-        committedFuture.complete(committedOffsets);
-        consumer.assign(partitions);
-        try (MockedConstruction<FetchCommittedOffsetsApplicationEvent> ignored = offsetFetchEventMocker(committedFuture)) {
-            // Poll with 250ms timeout to give the background thread time to process the events without timing out
-            consumer.poll(Duration.ofMillis(250));
-
-            verify(applicationEventHandler, atLeast(1))
-                .addAndGet(ArgumentMatchers.isA(ValidatePositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
-            verify(applicationEventHandler, atLeast(1))
+                .addAndGet(ArgumentMatchers.isA(FetchCommittedOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+            verify(applicationEventHandler, never())
+                .addAndGet(ArgumentMatchers.isA(ResetPositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+        } else {
+            // Verify there was not any FetchCommittedOffsets event but there should be a ResetPositions
+            verify(applicationEventHandler, never())
                 .addAndGet(ArgumentMatchers.isA(FetchCommittedOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
             verify(applicationEventHandler, atLeast(1))
                 .addAndGet(ArgumentMatchers.isA(ResetPositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
         }
     }
 
-    private void prepAutocommitOnClose() {
-        Node node = testBuilder.metadata.fetch().nodes().get(0);
-        testBuilder.client.prepareResponse(FindCoordinatorResponse.prepareResponse(Errors.NONE, "group-id", node));
-        if (!testBuilder.subscriptions.allConsumed().isEmpty()) {
-            List<TopicPartition> topicPartitions = new ArrayList<>(testBuilder.subscriptions.assignedPartitionsList());
-            testBuilder.client.prepareResponse(mockAutocommitResponse(
-                topicPartitions,
-                (short) 1,
-                Errors.NONE).responseBody());
-        }
+    private void testRefreshCommittedOffsetsSuccess(Set<TopicPartition> partitions,
+                                                    Map<TopicPartition, OffsetAndMetadata> committedOffsets) {
+        doReturn(committedOffsets)
+            .when(applicationEventHandler)
+            .addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+        doReturn(Fetch.empty()).when(fetchCollector).collectFetch(any(FetchBuffer.class));
+        doReturn(LeaderAndEpoch.noLeaderOrEpoch()).when(metadata).currentLeader(any());
+
+        consumer.assign(partitions);
+
+        // Poll with 250ms timeout to give the background thread time to process the events without timing out
+        consumer.poll(Duration.ofMillis(250));
+
+        verify(applicationEventHandler, atLeast(1))
+            .addAndGet(ArgumentMatchers.isA(ValidatePositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+        verify(applicationEventHandler, atLeast(1))
+            .addAndGet(ArgumentMatchers.isA(FetchCommittedOffsetsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
+        verify(applicationEventHandler, atLeast(1))
+            .addAndGet(ArgumentMatchers.isA(ResetPositionsApplicationEvent.class), ArgumentMatchers.isA(Timer.class));
     }
 
-    private ClientResponse mockAutocommitResponse(final List<TopicPartition> topicPartitions,
-                                                  final short apiKeyVersion,
-                                                  final Errors error) {
-        OffsetCommitResponseData responseData = new OffsetCommitResponseData();
-        List<OffsetCommitResponseData.OffsetCommitResponseTopic> responseTopics = new ArrayList<>();
-        topicPartitions.forEach(tp -> {
-            responseTopics.add(new OffsetCommitResponseData.OffsetCommitResponseTopic()
-                .setName(tp.topic())
-                .setPartitions(Collections.singletonList(
-                    new OffsetCommitResponseData.OffsetCommitResponsePartition()
-                        .setErrorCode(error.code())
-                        .setPartitionIndex(tp.partition()))));
-        });
-        responseData.setTopics(responseTopics);
-        OffsetCommitResponse response = mock(OffsetCommitResponse.class);
-        when(response.data()).thenReturn(responseData);
-        return new ClientResponse(
-            new RequestHeader(ApiKeys.OFFSET_COMMIT, apiKeyVersion, "", 1),
-            null,
-            "-1",
-            testBuilder.time.milliseconds(),
-            testBuilder.time.milliseconds(),
-            false,
-            null,
-            null,
-            new OffsetCommitResponse(responseData)
+    @Test
+    public void testSubscribeGeneratesEvent() {
+        consumer = setup();
+        String topic = "topic1";
+        consumer.subscribe(singletonList(topic));
+        assertEquals(singleton(topic), consumer.subscription());
+        assertTrue(consumer.assignment().isEmpty());
+        verify(applicationEventHandler).add(ArgumentMatchers.isA(SubscriptionChangeApplicationEvent.class));
+    }
+
+    @Test
+    public void testUnsubscribeGeneratesUnsubscribeEvent() {
+        consumer = setup();
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof UnsubscribeApplicationEvent);
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        consumer.unsubscribe();
+
+        assertTrue(consumer.subscription().isEmpty());
+        assertTrue(consumer.assignment().isEmpty());
+    }
+
+    @Test
+    public void testSubscribeToEmptyListActsAsUnsubscribe() {
+        consumer = setup();
+
+        Mockito.doAnswer(invocation -> {
+            CompletableApplicationEvent<?> event = invocation.getArgument(0);
+            assertTrue(event instanceof UnsubscribeApplicationEvent);
+            event.future().complete(null);
+            return null;
+        }).when(applicationEventHandler).add(any());
+
+        consumer.subscribe(Collections.emptyList());
+        assertTrue(consumer.subscription().isEmpty());
+        assertTrue(consumer.assignment().isEmpty());
+    }
+
+    @Test
+    public void testSubscribeToNullTopicCollection() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class, () -> consumer.subscribe((List<String>) null));
+    }
+
+    @Test
+    public void testSubscriptionOnNullTopic() {
+        consumer = setup();
+        assertThrows(IllegalArgumentException.class, () -> consumer.subscribe(singletonList(null)));
+    }
+
+    @Test
+    public void testSubscriptionOnEmptyTopic() {
+        consumer = setup();
+        String emptyTopic = "  ";
+        assertThrows(IllegalArgumentException.class, () -> consumer.subscribe(singletonList(emptyTopic)));
+    }
+
+    @Test
+    public void testGroupMetadataAfterCreationWithGroupIdIsNull() {
+        final Properties props = requiredConsumerProperties();
+        final ConsumerConfig config = new ConsumerConfig(props);
+        consumer = setup(config);
+
+        assertFalse(config.unused().contains(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG));
+        assertFalse(config.unused().contains(THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED));
+        final Throwable exception = assertThrows(InvalidGroupIdException.class, consumer::groupMetadata);
+        assertEquals(
+            "To use the group management or offset commit APIs, you must " +
+                "provide a valid " + ConsumerConfig.GROUP_ID_CONFIG + " in the consumer configuration.",
+            exception.getMessage()
         );
     }
+
+    @Test
+    public void testGroupMetadataAfterCreationWithGroupIdIsNotNull() {
+        final String groupId = "consumerGroupA";
+        final Properties props = requiredConsumerPropertiesAndGroupId("consumerGroupA");
+        final ConsumerConfig config = new ConsumerConfig(props);
+
+        consumer = setup(config);
+        final ConsumerGroupMetadata groupMetadata = consumer.groupMetadata();
+
+        assertEquals(groupId, groupMetadata.groupId());
+        assertEquals(Optional.empty(), groupMetadata.groupInstanceId());
+        assertEquals(JoinGroupRequest.UNKNOWN_GENERATION_ID, groupMetadata.generationId());
+        assertEquals(JoinGroupRequest.UNKNOWN_MEMBER_ID, groupMetadata.memberId());
+    }
+
+    @Test
+    public void testGroupMetadataAfterCreationWithGroupIdIsNotNullAndGroupInstanceIdSet() {
+        final String groupId = "consumerGroupA";
+        final String groupInstanceId = "groupInstanceId1";
+        final Properties props = requiredConsumerPropertiesAndGroupId(groupId);
+        props.put(ConsumerConfig.GROUP_INSTANCE_ID_CONFIG, groupInstanceId);
+        final ConsumerConfig config = new ConsumerConfig(props);
+        consumer = setup(config);
+
+        final ConsumerGroupMetadata groupMetadata = consumer.groupMetadata();
+
+        assertEquals(groupId, groupMetadata.groupId());
+        assertEquals(Optional.of(groupInstanceId), groupMetadata.groupInstanceId());
+        assertEquals(JoinGroupRequest.UNKNOWN_GENERATION_ID, groupMetadata.generationId());
+        assertEquals(JoinGroupRequest.UNKNOWN_MEMBER_ID, groupMetadata.memberId());
+    }
+
+    @Test
+    public void testGroupMetadataUpdateSingleCall() {
+        final String groupId = "consumerGroupA";
+        final ConsumerConfig config = new ConsumerConfig(requiredConsumerPropertiesAndGroupId(groupId));
+        consumer = setup(config);
+
+        doReturn(Fetch.empty()).when(fetchCollector).collectFetch(any(FetchBuffer.class));
+        doReturn(mkMap()).when(applicationEventHandler).addAndGet(any(FetchCommittedOffsetsApplicationEvent.class), any(Timer.class));
+
+        final int generation = 1;
+        final String memberId = "newMemberId";
+        final ConsumerGroupMetadata expectedGroupMetadata = new ConsumerGroupMetadata(
+            groupId,
+            generation,
+            memberId,
+            Optional.empty()
+        );
+        final GroupMetadataUpdateEvent groupMetadataUpdateEvent = new GroupMetadataUpdateEvent(
+            generation,
+            memberId
+        );
+        backgroundEventQueue.add(groupMetadataUpdateEvent);
+        consumer.assign(singletonList(new TopicPartition("topic", 0)));
+        consumer.poll(Duration.ZERO);
+
+        final ConsumerGroupMetadata actualGroupMetadata = consumer.groupMetadata();
+
+        assertEquals(expectedGroupMetadata, actualGroupMetadata);
+
+        final ConsumerGroupMetadata secondActualGroupMetadataWithoutUpdate = consumer.groupMetadata();
+
+        assertEquals(expectedGroupMetadata, secondActualGroupMetadataWithoutUpdate);
+    }
+
+    @Test
+    public void testBackgroundError() {
+        final String groupId = "consumerGroupA";
+        final ConsumerConfig config = new ConsumerConfig(requiredConsumerPropertiesAndGroupId(groupId));
+        consumer = setup(config);
+
+        final KafkaException expectedException = new KafkaException("Nobody expects the Spanish Inquisition");
+        final ErrorBackgroundEvent errorBackgroundEvent = new ErrorBackgroundEvent(expectedException);
+        backgroundEventQueue.add(errorBackgroundEvent);
+        consumer.assign(singletonList(new TopicPartition("topic", 0)));
+
+        final KafkaException exception = assertThrows(KafkaException.class, () -> consumer.poll(Duration.ZERO));
+
+        assertEquals(expectedException.getMessage(), exception.getMessage());
+    }
+
+    @Test
+    public void testMultipleBackgroundErrors() {
+        final String groupId = "consumerGroupA";
+        final ConsumerConfig config = new ConsumerConfig(requiredConsumerPropertiesAndGroupId(groupId));
+        consumer = setup(config);
+
+        final KafkaException expectedException1 = new KafkaException("Nobody expects the Spanish Inquisition");
+        final ErrorBackgroundEvent errorBackgroundEvent1 = new ErrorBackgroundEvent(expectedException1);
+        backgroundEventQueue.add(errorBackgroundEvent1);
+        final KafkaException expectedException2 = new KafkaException("Spam, Spam, Spam");
+        final ErrorBackgroundEvent errorBackgroundEvent2 = new ErrorBackgroundEvent(expectedException2);
+        backgroundEventQueue.add(errorBackgroundEvent2);
+        consumer.assign(singletonList(new TopicPartition("topic", 0)));
+
+        final KafkaException exception = assertThrows(KafkaException.class, () -> consumer.poll(Duration.ZERO));
+
+        assertEquals(expectedException1.getMessage(), exception.getMessage());
+        assertTrue(backgroundEventQueue.isEmpty());
+    }
+
+    @Test
+    public void testGroupRemoteAssignorUnusedIfGroupIdUndefined() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG, "someAssignor");
+        final ConsumerConfig config = new ConsumerConfig(props);
+        consumer = setup(config);
+
+        assertTrue(config.unused().contains(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG));
+    }
+
+    @Test
+    public void testGroupRemoteAssignorUnusedInGenericProtocol() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, "consumerGroupA");
+        props.put(ConsumerConfig.GROUP_PROTOCOL_CONFIG, GroupProtocol.GENERIC.name().toLowerCase(Locale.ROOT));
+        props.put(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG, "someAssignor");
+        final ConsumerConfig config = new ConsumerConfig(props);
+        consumer = setup(config);
+
+        assertTrue(config.unused().contains(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG));
+    }
+
+    @Test
+    public void testGroupRemoteAssignorUsedInConsumerProtocol() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, "consumerGroupA");
+        props.put(ConsumerConfig.GROUP_PROTOCOL_CONFIG, GroupProtocol.CONSUMER.name().toLowerCase(Locale.ROOT));
+        props.put(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG, "someAssignor");
+        final ConsumerConfig config = new ConsumerConfig(props);
+        consumer = setup(config);
+
+        assertFalse(config.unused().contains(ConsumerConfig.GROUP_REMOTE_ASSIGNOR_CONFIG));
+    }
+
+    @Test
+    public void testGroupIdNull() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, 10000);
+        props.put(THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED, true);
+        final ConsumerConfig config = new ConsumerConfig(props);
+
+        consumer = setup(config);
+        assertFalse(config.unused().contains(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG));
+        assertFalse(config.unused().contains(THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED));
+    }
+
+    @Test
+    public void testGroupIdNotNullAndValid() {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, "consumerGroupA");
+        props.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, 10000);
+        props.put(THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED, true);
+        final ConsumerConfig config = new ConsumerConfig(props);
+
+        consumer = setup(config);
+        assertTrue(config.unused().contains(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG));
+        assertTrue(config.unused().contains(THROW_ON_FETCH_STABLE_OFFSET_UNSUPPORTED));
+    }
+
+    @Test
+    public void testGroupIdEmpty() {
+        testInvalidGroupId("");
+    }
+
+    @Test
+    public void testGroupIdOnlyWhitespaces() {
+        testInvalidGroupId("       ");
+    }
+
+    private void testInvalidGroupId(final String groupId) {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
+        final ConsumerConfig config = new ConsumerConfig(props);
+
+        final Exception exception = assertThrows(
+            KafkaException.class,
+            () -> consumer = setup(config)
+        );
+
+        assertEquals("Failed to construct kafka consumer", exception.getMessage());
+    }
+
+    private Properties requiredConsumerPropertiesAndGroupId(final String groupId) {
+        final Properties props = requiredConsumerProperties();
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
+        return props;
+    }
+
+    private Properties requiredConsumerProperties() {
+        final Properties props = new Properties();
+        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
+        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
+        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9091");
+        return props;
+    }
+
+    @Test
+    public void testLongPollWaitIsLimited() {
+        consumer = setup();
+        String topicName = "topic1";
+        consumer.subscribe(singletonList(topicName));
+
+        assertEquals(singleton(topicName), consumer.subscription());
+        assertTrue(consumer.assignment().isEmpty());
+
+        final int partition = 3;
+        final TopicPartition tp = new TopicPartition(topicName, partition);
+        final List<ConsumerRecord<String, String>> records = asList(
+            new ConsumerRecord<>(topicName, partition, 2, "key1", "value1"),
+            new ConsumerRecord<>(topicName, partition, 3, "key2", "value2")
+        );
+
+        // On the first iteration, return no data; on the second, return two records
+        doAnswer(invocation -> {
+            // Mock the subscription being assigned as the first fetch is collected
+            consumer.subscriptions().assignFromSubscribed(Collections.singleton(tp));
+            return Fetch.empty();
+        })
+            .doAnswer(invocation -> Fetch.forPartition(tp, records, true))
+            .when(fetchCollector).collectFetch(any(FetchBuffer.class));
+
+        // And then poll for up to 10000ms, which should return 2 records without timing out
+        ConsumerRecords<?, ?> returnedRecords = consumer.poll(Duration.ofMillis(10000));
+        assertEquals(2, returnedRecords.count());
+
+        assertEquals(singleton(topicName), consumer.subscription());
+        assertEquals(singleton(tp), consumer.assignment());
+    }
+
+
+    private HashMap<TopicPartition, OffsetAndMetadata> mockTopicPartitionOffset() {
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, OffsetAndMetadata> topicPartitionOffsets = new HashMap<>();
+        topicPartitionOffsets.put(t0, new OffsetAndMetadata(10L));
+        topicPartitionOffsets.put(t1, new OffsetAndMetadata(20L));
+        return topicPartitionOffsets;
+    }
+
+    private HashMap<TopicPartition, OffsetAndTimestamp> mockOffsetAndTimestamp() {
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, OffsetAndTimestamp> offsetAndTimestamp = new HashMap<>();
+        offsetAndTimestamp.put(t0, new OffsetAndTimestamp(5L, 1L));
+        offsetAndTimestamp.put(t1, new OffsetAndTimestamp(6L, 3L));
+        return offsetAndTimestamp;
+    }
+
+    private HashMap<TopicPartition, Long> mockTimestampToSearch() {
+        final TopicPartition t0 = new TopicPartition("t0", 2);
+        final TopicPartition t1 = new TopicPartition("t0", 3);
+        HashMap<TopicPartition, Long> timestampToSearch = new HashMap<>();
+        timestampToSearch.put(t0, 1L);
+        timestampToSearch.put(t1, 2L);
+        return timestampToSearch;
+    }
+
 }
